refactor: decisions module refactor
