+++ src/decisions/decisions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { DecisionsController } from './controller/decisions/decisions.controller';
-import { DecisionsService } from './services/decisions/decisions.service';
+import { DecisionsController } from './controller/decisions.controller';
+import { DecisionsService } from './services/decisions.service';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
